fix all test; add function wrapper instead of using aop;

+++ src/test/java/io/transwarp/generate/common/TableUtilTest.java
@@ -16,7 +16,7 @@ public class TableUtilTest {
   @Test
   public void randomCol() throws Exception {
     final Table table = DDLParserTest.getTable();
-    assertTrue(TableUtil.sameTypeRandomCol(table, DataType.INT).isPresent());
+    assertTrue(TableUtil.sameTypeRandomCol(table, DataType.DECIMAL).isPresent());
     assertTrue(TableUtil.sameTypeRandomCol(table, DataType.DATE).isPresent());
     assertTrue(TableUtil.sameTypeRandomCol(table, SequenceDataType.CHARS).isPresent());
     assertTrue(!TableUtil.sameTypeRandomCol(table, DataType.BOOL).isPresent());
